// src/app/model/project.model.ts
  export class ProjectDetails {
    constructor() {}

    projectTitle?: string;
    projectDescription?: string;
    estimatedValue?: string;
    startDate?: string;
    endDate?: string;
    status?: string;
    projectManager?: string;
    projectManagerEmail?: string;
    technicalLeader?: string;
    technicalLeaderEmail?: string;
    unit?: string;
    projectCategory?: string;
    inHouse?: string;
    comments?: string;
    ptsRef?: string;
    id?: number;
    dateCreated?: string;

    otherUnit?: string;
    outSourced?: string;

    archived?: string;
    completed?: string;

    totalDays?: number;
    reason?: string;
  }

  